refactored button component inside article to stop error occurring
