+++ src/App.jsx
@@ -8,7 +8,6 @@ import ItemsPerPage from "./components/ItemsPerPage";
 import { Routes, Route } from "react-router-dom";
 import SingleArticle from "./components/SingleArticle";
 import Comments from "./components/Comments";
-import VoteButton from "./components/VoteButton";
 
 function App() {
   const [pageNumber, setPageNumber] = useState(1);
@@ -47,7 +46,6 @@ function App() {
           path="/:article_id"
           element={
             <>
-              {/* <VoteButton /> */}
               <div className="articles-comments-container">
                 <SingleArticle />
                 <Comments />
